+++ qrcode/main.py
@@ -186,8 +186,8 @@ class QRCode:
                 from qrcode.image.pil import PilImage
                 image_factory = PilImage
 
-        im = image_factory(self.border, self.modules_count, self.box_size,
-                **kwargs)
+        im = image_factory(
+            self.border, self.modules_count, self.box_size, **kwargs)
         for r in range(self.modules_count):
             for c in range(self.modules_count):
                 if self.modules[r][c]: